Extract protocol name generation into public module
